style(limits): drop five no-op u8 casts left by the opcode port
`DigMessage::msg_type` is already a raw `u8` since the port off
`ProtocolMessageTypes`, so these `as u8` casts are no-ops that trip
`clippy::unnecessary_cast` under `-D warnings`.

+++ src/connection/inbound_limits.rs
@@ -180,7 +180,7 @@ pub(crate) fn rejected_frame_incurs_penalty(msg: &DigMessage) -> bool {
 /// `false` (unreachable for 221/222 — the completeness guard pins their rows).
 fn exceeds_dig_wire_max_size(msg: &DigMessage) -> bool {
     dig_extension_rate_limits_map()
-        .get(&(msg.msg_type as u8))
+        .get(&msg.msg_type)
         .map(|row| (msg.data.len() as f64) > row.max_size)
         .unwrap_or(false)
 }
